+++ src/crawlee/browsers/__init__.py
@@ -1,7 +1,14 @@
+# ruff: noqa: E402, TID252
+
 from crawlee._utils.try_import import install_import_hook as _install_import_hook
 from crawlee._utils.try_import import try_import as _try_import
 
 _install_import_hook(__name__)
+
+# Due to patch_browserforge
+from .._browserforge_workaround import patch_browserforge
+
+patch_browserforge()
 
 # The following imports are wrapped in try_import to handle optional dependencies,
 # ensuring the module can still function even if these dependencies are missing.