+++ src/crawlee/fingerprint_suite/__init__.py
@@ -1,3 +1,10 @@
+# ruff: noqa: E402, TID252
+
+# Due to patch_browserforge
+from .._browserforge_workaround import patch_browserforge
+
+patch_browserforge()
+
 from ._browserforge_adapter import BrowserforgeFingerprintGenerator as DefaultFingerprintGenerator
 from ._fingerprint_generator import FingerprintGenerator
 from ._header_generator import HeaderGenerator